Power BI
This project is related to "Madhav Store", an E-Commerce store. In the Excel sheet which contains the sales related data of the store. This is the complete Interactive-led dashboards. Slicers makes it more fun to use and analyze the data properly. Used multiple charts to create this project. It was the most amazing project which I have worked on till date. And the output is mesmerizing, watch it yourself. 

I've done;
1. Data Extraction
2. Data Transformation
3. Data visualization &
4. Data Analysis.