Microsoft Excel
Excel Interactive Dashboards using Pivot tables and Slicers:
This is one of the largest data sets which I've worked on Excel with more than 30,000 columns. This is related to "Vrinda Store Annual Report - 2022" which you can see all the data and the dashboards which I've created using PIVOT and Slicer options. This is like a complete Interactive-led dashboards similar to Power BI, but created with Microsoft Excel.
I've done;
1. Data Cleaning
2. Data Filtering / sorting with Pivot Tables
3. Data visualization - graphs, charts, slicers
4. Data Analysis - from the insights of multiple dashboards